% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreNLPsetup-package.R
\docType{package}
\name{coreNLPsetup}
\alias{coreNLPsetup}
\alias{package-coreNLPsetup}
\alias{coreNLPsetup-package}
\title{A Set of Tools to Ensure Proper Setup of Stanford's \pkg{CoreNLP}}
\description{
Ensures that CoreNLP is installed and is playing nicely with \pkg{Java}.
}
